requirements.txt
django==1.11.3
coreapi==2.3.1
markdown==2.6.8
django-suit==0.2.25
django-filter==1.0.4
django-guardian==1.4.9
django-crispy-forms==1.6.1
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0